[minor] Support OCP node configuration
Adds support for applying labels and taints to nodes, identifying nodes either by name or their index in the list of nodes.

Also updates the db2 role to seperate the task of applying labels and taints from the role.  Users that wish to run with a dedicated Db2 node should use the new `ocp_node_config` role before running the `db2` role.  This will ensure that the node is correctly set up before starting the installation pipeline, instead of - as it works today - the node being tainted and drained midway through a MAS install.