[patch] Remove redhat_connect_username/password
We needed the redhat_connect_username  and redhat_connect_password for mirroring UDS images because some images were in the RedHat container registry instead of the IBM container registry.  Now that we have completed the migration to DRO there is no need for this anymore/mirroring UDS is no longer supported as UDS has been shut down.