[patch] Always set failed=False
We don't want to fail if we can't find any default storage classes, doing so will result in roles/playbooks failing in environments where none of the default storage classes are available.  We use the success=false to track when we couldn't find a default storage class, which does not trigger Ansible treating the action as failed.