[major] Remove deprecated UDS role (and migration support)
It has been a long time since UDS EOL (Spring 2024).  It is time to remove the role from the collection and clean up any lingering references to UDS and the complexity continuing to have those in our pipelines creates.  Any customers still running UDS will need to update to an older catalog using an older version of the collection/CLI, or handle the migration independently.